Move code around (Fixes #113)
- Move all variants into src/variants
- Move shared variant code into src/lib
- Move tests into __tests__ for each of the above directories
