--- packages/shared/src/baduk/index.ts
@@ -1,1 +1,0 @@
-export * from "./baduk";
