--- packages/shared/src/capture/index.ts
@@ -1,1 +1,0 @@
-export * from "./capture";
