--- packages/shared/src/chess/index.ts
@@ -1,1 +1,0 @@
-export { ChessGame } from "./chess";
