+++ packages/shared/src/index.ts
@@ -1,12 +1,12 @@
-export * from "./baduk";
-export * from "./abstractAlternatingOnGrid";
-export * from "./phantom";
-export * from "./capture";
-export * from "./parallel";
+export * from "./variants/baduk";
+export * from "./lib/abstractAlternatingOnGrid";
+export * from "./variants/phantom";
+export * from "./variants/capture";
+export * from "./variants/parallel";
 export * from "./abstract_game";
 export * from "./game_map";
 export * from "./api_types";
-export * from "./grid";
-export * from "./coordinate";
+export * from "./lib/grid";
+export * from "./lib/coordinate";
 export * from "./time_control";
-export * from "./utils";
+export * from "./lib/utils";
