+++ packages/shared/src/lib/__tests__/ko_detector.test.ts
@@ -1,4 +1,4 @@
-import { SuperKoDetector, KoError } from "./ko_detector";
+import { SuperKoDetector, KoError } from "../ko_detector";
 
 test("superko detector throws if same state pushed twice", () => {
   const detector = new SuperKoDetector();
