+++ packages/shared/src/lib/abstractAlternatingOnGrid/index.ts
