+++ packages/shared/src/lib/coordinate.ts
