+++ packages/shared/src/lib/ko_detector.ts
