--- packages/shared/src/parallel/index.ts
@@ -1,1 +1,0 @@
-export * from "./parallel";
