--- packages/shared/src/phantom/index.ts
@@ -1,1 +1,0 @@
-export * from "./phantom";
