+++ packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/Interfaces/IBadukBoard.ts
