+++ packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/PolygonalBoardHelper.ts
