+++ packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/StoneState.ts
