+++ packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/intersection.ts
