+++ packages/shared/src/variants/badukWithAbstractBoard/index.ts
@@ -1,0 +1,1 @@
+export * from "./badukWithAbstractBoard";
