+++ packages/vue-client/src/components/GameCreation/BadukWithAbstractBoardConfigForm.vue
@@ -2,7 +2,7 @@
 import {
   BoardPattern,
   type BadukWithAbstractBoardConfig,
-} from "@ogfcommunity/variants-shared/src/badukWithAbstractBoard/badukWithAbstractBoard";
+} from "@ogfcommunity/variants-shared/src/variants/badukWithAbstractBoard";
 import { type Ref, ref, watch } from "vue";
 
 const props = defineProps<{ initialConfig: BadukWithAbstractBoardConfig }>();
