+++ packages/vue-client/src/components/boards/BadukBoardAbstract.vue
@@ -3,7 +3,7 @@ import { Color } from "@ogfcommunity/variants-shared";
 import type {
   BadukWithAbstractBoardConfig,
   BadukWithAbstractBoardState,
-} from "@ogfcommunity/variants-shared/src/badukWithAbstractBoard/badukWithAbstractBoard";
+} from "@ogfcommunity/variants-shared/src/variants/badukWithAbstractBoard";
 import { computed } from "vue";
 
 const props = defineProps<{
